Enhance dep-relations for paasng.engine, part 1
