+++ apiserver/paasng/paasng/extensions/declarative/deployment/validations.py
@@ -20,9 +20,9 @@ import cattr
 from rest_framework import serializers
 
 from paasng.engine.constants import ConfigVarEnvName
-from paasng.engine.serializers import field_env_var_key
 from paasng.extensions.declarative.deployment.resources import DeploymentDesc
 from paasng.extensions.declarative.serializers import validate_language
+from paasng.utils.serializers import field_env_var_key
 
 
 class EnvVariableSLZ(serializers.Serializer):
